Rename code review groups components
