+++ apps/src/templates/codeReviewGroups/StudentGroup.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {Droppable} from 'react-beautiful-dnd';
-import CodeReviewGroupMember from './CodeReviewGroupMember';
+import Student from './Student';
 
 // A CodeReviewGroup is a component that
 // CodeReviewGroupMembers can be dragged between as teachers are arranging students
@@ -9,7 +9,7 @@ import CodeReviewGroupMember from './CodeReviewGroupMember';
 // These are called "Droppables" in the package we're using (React Beautiful DnD).
 // More information on React Beautiful DnD can be found here:
 // https://github.com/atlassian/react-beautiful-dnd
-export default function CodeReviewGroup({droppableId, members}) {
+export default function StudentGroup({droppableId, members}) {
   return (
     <Droppable key={droppableId} droppableId={droppableId}>
       {(provided, snapshot) => (
@@ -19,7 +19,7 @@ export default function CodeReviewGroup({droppableId, members}) {
           {...provided.droppableProps}
         >
           {members.map((member, index) => (
-            <CodeReviewGroupMember
+            <Student
               followerId={member.followerId}
               name={member.name}
               index={index}
@@ -35,7 +35,7 @@ export default function CodeReviewGroup({droppableId, members}) {
 
 // Each group needs a unique droppableId (rather than a database-provided group ID)
 // so that we can create new groups on the fly without any interaction with our backend.
-CodeReviewGroup.propTypes = {
+StudentGroup.propTypes = {
   droppableId: PropTypes.string.isRequired,
   members: PropTypes.array.isRequired
 };
